Move user search into mixin
